+++ xroad-catalog-persistence/src/main/java/fi/vrk/xroad/catalog/persistence/dto/DescriptorInfoList.java
@@ -22,7 +22,6 @@
  */
 package fi.vrk.xroad.catalog.persistence.dto;
 
-import fi.vrk.xroad.catalog.persistence.entity.ErrorLog;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -39,15 +38,9 @@ import java.util.List;
 @EqualsAndHashCode
 @Getter
 @Setter
-public class ErrorLogResponse implements Serializable {
+public class DescriptorInfoList implements Serializable {
 
-    private static final long serialVersionUID = 4049861366378846285L;
+    private static final long serialVersionUID = 4049561366368846285L;
 
-    private Integer pageNumber;
-
-    private Integer pageSize;
-
-    private Integer numberOfPages;
-
-    private List<ErrorLog> errorLogList;
+    private List<DescriptorInfo> descriptorInfoList;
 }